support artifact metadata
