+++ backend/Makefile
@@ -5,7 +5,7 @@ CSV_PATH=backend/third_party_licenses
 # Whenever build command for any of the binaries change, we should update them both here and in backend/Dockerfiles.
 
 .PHONY: all
-all: license_apiserver license_persistence_agent license_cache_server license_swf license_viewer
+all: license_apiserver license_persistence_agent license_cache_server license_swf license_viewer license_launcher license_driver
 
 .PHONY: clean
 clean:
@@ -21,6 +21,10 @@ $(BUILD)/swf:
 	GO111MODULE=on go build -o $(BUILD)/swf github.com/kubeflow/pipelines/backend/src/crd/controller/scheduledworkflow
 $(BUILD)/viewer:
 	GO111MODULE=on go build -o $(BUILD)/viewer github.com/kubeflow/pipelines/backend/src/crd/controller/viewer
+$(BUILD)/launcher:
+	GO111MODULE=on go build -o $(BUILD)/launcher github.com/kubeflow/pipelines/backend/src/v2/component
+$(BUILD)/driver:
+	GO111MODULE=on go build -o $(BUILD)/driver github.com/kubeflow/pipelines/backend/src/v2/driver
 
 # Update licenses info after dependencies changed.
 # See README.md#updating-licenses-info section for more details.
@@ -39,6 +43,13 @@ license_swf: $(BUILD)/swf
 .PHONY: license_viewer
 license_viewer: $(BUILD)/viewer
 	cd $(MOD_ROOT) && go-licenses csv ./backend/src/crd/controller/viewer > $(CSV_PATH)/viewer.csv
+.PHONY: license_launcher
+license_launcher: $(BUILD)/launcher
+	cd $(MOD_ROOT) && go-licenses csv ./backend/src/v2/component > $(CSV_PATH)/launcher.csv
+.PHONY: license_driver
+license_driver: $(BUILD)/driver
+	cd $(MOD_ROOT) && go-licenses csv ./backend/src/v2/driver > $(CSV_PATH)/driver.csv
+
 
 .PHONY: image_all
 image_all: image_apiserver image_persistence_agent image_cache image_swf image_viewer image_visualization
