--- backend/src/v2/component/constants.go
@@ -1,28 +1,0 @@
-// Copyright 2021 The Kubeflow Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package component
-
-const (
-	VolumePathKFPLauncher = "/kfp-launcher"
-	KFPLauncherPath       = VolumePathKFPLauncher + "/launch"
-
-	// Env var names
-	EnvPodName = "KFP_POD_NAME"
-	EnvPodUID  = "KFP_POD_UID"
-
-	// Env vars in metadata-grpc-configmap
-	EnvMetadataHost = "METADATA_GRPC_SERVICE_HOST"
-	EnvMetadataPort = "METADATA_GRPC_SERVICE_PORT"
-)
