+++ backend/src/v2/util/util.go
@@ -1,0 +1,341 @@
+// Copyright 2021-2023 The Kubeflow Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//	https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+package util
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/golang/glog"
+	"github.com/kubeflow/pipelines/api/v2alpha1/go/pipelinespec"
+	"github.com/kubeflow/pipelines/backend/src/v2/expression"
+	"github.com/kubeflow/pipelines/backend/src/v2/metadata"
+	"google.golang.org/protobuf/types/known/structpb"
+)
+
+const (
+	OutputMetadataFilepath = "/tmp/kfp_outputs/output_metadata.json"
+
+	VolumePathKFPLauncher = "/kfp-launcher"
+	KFPLauncherPath       = VolumePathKFPLauncher + "/launch"
+
+	// Env var names
+	EnvPodName = "KFP_POD_NAME"
+	EnvPodUID  = "KFP_POD_UID"
+
+	// Env vars in metadata-grpc-configmap
+	EnvMetadataHost = "METADATA_GRPC_SERVICE_HOST"
+	EnvMetadataPort = "METADATA_GRPC_SERVICE_PORT"
+)
+
+func ResolveInputs(ctx context.Context, dag *metadata.DAG, iterationIndex *int, pipeline *metadata.Pipeline, task *pipelinespec.PipelineTaskSpec, inputsSpec *pipelinespec.ComponentInputsSpec, mlmd *metadata.Client, expr *expression.Expr) (inputs *pipelinespec.ExecutorInput_Inputs, err error) {
+	defer func() {
+		if err != nil {
+			err = fmt.Errorf("failed to resolve inputs: %w", err)
+		}
+	}()
+	inputParams, _, err := dag.Execution.GetParameters()
+	if err != nil {
+		return nil, err
+	}
+	glog.Infof("parent DAG input parameters %+v", inputParams)
+	inputs = &pipelinespec.ExecutorInput_Inputs{
+		ParameterValues: make(map[string]*structpb.Value),
+		Artifacts:       make(map[string]*pipelinespec.ArtifactList),
+	}
+	isIterationDriver := iterationIndex != nil
+
+	handleParameterExpressionSelector := func() error {
+		for name, paramSpec := range task.GetInputs().GetParameters() {
+			var selector string
+			if selector = paramSpec.GetParameterExpressionSelector(); selector == "" {
+				continue
+			}
+			wrap := func(e error) error {
+				return fmt.Errorf("resolving parameter %q: evaluation of parameter expression selector %q failed: %w", name, selector, e)
+			}
+			value, ok := inputs.ParameterValues[name]
+			if !ok {
+				return wrap(fmt.Errorf("value not found in inputs"))
+			}
+			selected, err := expr.Select(value, selector)
+			if err != nil {
+				return wrap(err)
+			}
+			inputs.ParameterValues[name] = selected
+		}
+		return nil
+	}
+	handleParamTypeValidationAndConversion := func() error {
+		// TODO(Bobgy): verify whether there are inputs not in the inputs spec.
+		for name, spec := range inputsSpec.GetParameters() {
+			if task.GetParameterIterator() != nil {
+				if !isIterationDriver && task.GetParameterIterator().GetItemInput() == name {
+					// It's expected that an iterator does not have iteration item input parameter,
+					// because only iterations get the item input parameter.
+					continue
+				}
+				if isIterationDriver && task.GetParameterIterator().GetItems().GetInputParameter() == name {
+					// It's expected that an iteration does not have iteration items input parameter,
+					// because only the iterator has it.
+					continue
+				}
+			}
+			value, hasValue := inputs.GetParameterValues()[name]
+
+			// Handle when parameter does not have input value
+			if !hasValue && !inputsSpec.GetParameters()[name].GetIsOptional() {
+				// When parameter is not optional and there is no input value, first check if there is a default value,
+				// if there is a default value, use it as the value of the parameter.
+				// if there is no default value, report error.
+				if inputsSpec.GetParameters()[name].GetDefaultValue() == nil {
+					return fmt.Errorf("neither value nor default value provided for non-optional parameter %q", name)
+				}
+			} else if !hasValue && inputsSpec.GetParameters()[name].GetIsOptional() {
+				// When parameter is optional and there is no input value, value comes from default value.
+				// But we don't pass the default value here. They are resolved internally within the component.
+				// Note: in the past the backend passed the default values into the component. This is a behavior change.
+				// See discussion: https://github.com/kubeflow/pipelines/pull/8765#discussion_r1119477085
+				continue
+			}
+
+			switch spec.GetParameterType() {
+			case pipelinespec.ParameterType_STRING:
+				_, isValueString := value.GetKind().(*structpb.Value_StringValue)
+				if !isValueString {
+					// TODO(Bobgy): discuss whether we want to allow auto type conversion
+					// all parameter types can be consumed as JSON string
+					text, err := metadata.PbValueToText(value)
+					if err != nil {
+						return fmt.Errorf("converting input parameter %q to string: %w", name, err)
+					}
+					inputs.GetParameterValues()[name] = structpb.NewStringValue(text)
+				}
+			default:
+				typeMismatch := func(actual string) error {
+					return fmt.Errorf("input parameter %q type mismatch: expect %s, got %s", name, spec.GetParameterType(), actual)
+				}
+				switch v := value.GetKind().(type) {
+				case *structpb.Value_NullValue:
+					return fmt.Errorf("got null for input parameter %q", name)
+				case *structpb.Value_StringValue:
+					// TODO(Bobgy): consider whether we support parsing string as JSON for any other types.
+					if spec.GetParameterType() != pipelinespec.ParameterType_STRING {
+						return typeMismatch("string")
+					}
+				case *structpb.Value_NumberValue:
+					if spec.GetParameterType() != pipelinespec.ParameterType_NUMBER_DOUBLE && spec.GetParameterType() != pipelinespec.ParameterType_NUMBER_INTEGER {
+						return typeMismatch("number")
+					}
+				case *structpb.Value_BoolValue:
+					if spec.GetParameterType() != pipelinespec.ParameterType_BOOLEAN {
+						return typeMismatch("bool")
+					}
+				case *structpb.Value_ListValue:
+					if spec.GetParameterType() != pipelinespec.ParameterType_LIST {
+						return typeMismatch("list")
+					}
+				case *structpb.Value_StructValue:
+					if spec.GetParameterType() != pipelinespec.ParameterType_STRUCT {
+						return typeMismatch("struct")
+					}
+				default:
+					return fmt.Errorf("parameter %s has unknown protobuf.Value type: %T", name, v)
+				}
+			}
+		}
+		return nil
+	}
+	// this function has many branches, so it's hard to add more postprocess steps
+	// TODO(Bobgy): consider splitting this function into several sub functions
+	defer func() {
+		if err == nil {
+			err = handleParameterExpressionSelector()
+		}
+		if err == nil {
+			err = handleParamTypeValidationAndConversion()
+		}
+	}()
+	// resolve input parameters
+	if isIterationDriver {
+		// resolve inputs for iteration driver is very different
+		artifacts, err := mlmd.GetInputArtifactsByExecutionID(ctx, dag.Execution.GetID())
+		if err != nil {
+			return nil, err
+		}
+		inputs.ParameterValues = inputParams
+		inputs.Artifacts = artifacts
+		switch {
+		case task.GetArtifactIterator() != nil:
+			return nil, fmt.Errorf("artifact iterator not implemented yet")
+		case task.GetParameterIterator() != nil:
+			var itemsInput string
+			if task.GetParameterIterator().GetItems().GetInputParameter() != "" {
+				// input comes from outside the component
+				itemsInput = task.GetParameterIterator().GetItems().GetInputParameter()
+			} else if task.GetParameterIterator().GetItemInput() != "" {
+				// input comes from static input
+				itemsInput = task.GetParameterIterator().GetItemInput()
+			} else {
+				return nil, fmt.Errorf("cannot retrieve parameter iterator")
+			}
+			items, err := GetItems(inputs.ParameterValues[itemsInput])
+			if err != nil {
+				return nil, err
+			}
+			if *iterationIndex >= len(items) {
+				return nil, fmt.Errorf("bug: %v items found, but getting index %v", len(items), *iterationIndex)
+			}
+			delete(inputs.ParameterValues, itemsInput)
+			inputs.ParameterValues[task.GetParameterIterator().GetItemInput()] = items[*iterationIndex]
+		default:
+			return nil, fmt.Errorf("bug: iteration_index>=0, but task iterator is empty")
+		}
+		return inputs, nil
+	}
+	// get executions in context on demand
+	var tasksCache map[string]*metadata.Execution
+	getDAGTasks := func() (map[string]*metadata.Execution, error) {
+		if tasksCache != nil {
+			return tasksCache, nil
+		}
+		tasks, err := mlmd.GetExecutionsInDAG(ctx, dag, pipeline)
+		if err != nil {
+			return nil, err
+		}
+		tasksCache = tasks
+		return tasks, nil
+	}
+	for name, paramSpec := range task.GetInputs().GetParameters() {
+		paramError := func(err error) error {
+			return fmt.Errorf("resolving input parameter %s with spec %s: %w", name, paramSpec, err)
+		}
+		switch t := paramSpec.Kind.(type) {
+		case *pipelinespec.TaskInputsSpec_InputParameterSpec_ComponentInputParameter:
+			componentInput := paramSpec.GetComponentInputParameter()
+			if componentInput == "" {
+				return nil, paramError(fmt.Errorf("empty component input"))
+			}
+			v, ok := inputParams[componentInput]
+			if !ok {
+				return nil, paramError(fmt.Errorf("parent DAG does not have input parameter %s", componentInput))
+			}
+			inputs.ParameterValues[name] = v
+
+		case *pipelinespec.TaskInputsSpec_InputParameterSpec_TaskOutputParameter:
+			taskOutput := paramSpec.GetTaskOutputParameter()
+			if taskOutput.GetProducerTask() == "" {
+				return nil, paramError(fmt.Errorf("producer task is empty"))
+			}
+			if taskOutput.GetOutputParameterKey() == "" {
+				return nil, paramError(fmt.Errorf("output parameter key is empty"))
+			}
+			tasks, err := getDAGTasks()
+			if err != nil {
+				return nil, paramError(err)
+			}
+			producer, ok := tasks[taskOutput.GetProducerTask()]
+			if !ok {
+				return nil, paramError(fmt.Errorf("cannot find producer task %q", taskOutput.GetProducerTask()))
+			}
+			_, outputs, err := producer.GetParameters()
+			if err != nil {
+				return nil, paramError(fmt.Errorf("get producer output parameters: %w", err))
+			}
+			param, ok := outputs[taskOutput.GetOutputParameterKey()]
+			if !ok {
+				return nil, paramError(fmt.Errorf("cannot find output parameter key %q in producer task %q", taskOutput.GetOutputParameterKey(), taskOutput.GetProducerTask()))
+			}
+			inputs.ParameterValues[name] = param
+		case *pipelinespec.TaskInputsSpec_InputParameterSpec_RuntimeValue:
+			runtimeValue := paramSpec.GetRuntimeValue()
+			switch t := runtimeValue.Value.(type) {
+			case *pipelinespec.ValueOrRuntimeParameter_Constant:
+				inputs.ParameterValues[name] = runtimeValue.GetConstant()
+			default:
+				return nil, paramError(fmt.Errorf("param runtime value spec of type %T not implemented", t))
+			}
+
+		// TODO(Bobgy): implement the following cases
+		// case *pipelinespec.TaskInputsSpec_InputParameterSpec_TaskFinalStatus_:
+		default:
+			return nil, paramError(fmt.Errorf("parameter spec of type %T not implemented yet", t))
+		}
+	}
+	for name, artifactSpec := range task.GetInputs().GetArtifacts() {
+		artifactError := func(err error) error {
+			return fmt.Errorf("failed to resolve input artifact %s with spec %s: %w", name, artifactSpec, err)
+		}
+		switch t := artifactSpec.Kind.(type) {
+		case *pipelinespec.TaskInputsSpec_InputArtifactSpec_ComponentInputArtifact:
+			return nil, artifactError(fmt.Errorf("component input artifact not implemented yet"))
+
+		case *pipelinespec.TaskInputsSpec_InputArtifactSpec_TaskOutputArtifact:
+			taskOutput := artifactSpec.GetTaskOutputArtifact()
+			if taskOutput.GetProducerTask() == "" {
+				return nil, artifactError(fmt.Errorf("producer task is empty"))
+			}
+			if taskOutput.GetOutputArtifactKey() == "" {
+				return nil, artifactError(fmt.Errorf("output artifact key is empty"))
+			}
+			tasks, err := getDAGTasks()
+			if err != nil {
+				return nil, artifactError(err)
+			}
+			producer, ok := tasks[taskOutput.GetProducerTask()]
+			if !ok {
+				return nil, artifactError(fmt.Errorf("cannot find producer task %q", taskOutput.GetProducerTask()))
+			}
+			// TODO(Bobgy): cache results
+			outputs, err := mlmd.GetOutputArtifactsByExecutionId(ctx, producer.GetID())
+			if err != nil {
+				return nil, artifactError(err)
+			}
+			artifact, ok := outputs[taskOutput.GetOutputArtifactKey()]
+			if !ok {
+				return nil, artifactError(fmt.Errorf("cannot find output artifact key %q in producer task %q", taskOutput.GetOutputArtifactKey(), taskOutput.GetProducerTask()))
+			}
+			runtimeArtifact, err := artifact.ToRuntimeArtifact()
+			if err != nil {
+				return nil, artifactError(err)
+			}
+			inputs.Artifacts[name] = &pipelinespec.ArtifactList{
+				Artifacts: []*pipelinespec.RuntimeArtifact{runtimeArtifact},
+			}
+		default:
+			return nil, artifactError(fmt.Errorf("artifact spec of type %T not implemented yet", t))
+		}
+	}
+	// TODO(Bobgy): validate executor inputs match component inputs definition
+	return inputs, nil
+}
+
+// Get iteration items from a structpb.Value.
+// Return value may be
+// * a list of JSON serializable structs
+// * a list of structpb.Value
+func GetItems(value *structpb.Value) (items []*structpb.Value, err error) {
+	switch v := value.GetKind().(type) {
+	case *structpb.Value_ListValue:
+		return v.ListValue.GetValues(), nil
+	case *structpb.Value_StringValue:
+		listValue := structpb.Value{}
+		if err = listValue.UnmarshalJSON([]byte(v.StringValue)); err != nil {
+			return nil, err
+		}
+		return listValue.GetListValue().GetValues(), nil
+	default:
+		return nil, fmt.Errorf("value of type %T cannot be iterated", v)
+	}
+}
